# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "3D Print API",
    "version": "1.0.0",
    "description": "API documentation for W05 Final Project Part"
  },
  "servers": [
    {
      "url": "https://w05-final-project-part.onrender.com",
      "description": "Render deployment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/products": {
      "get": { "summary": "Get all products", "responses": { "200": { "description": "List of products" } } },
      "post": {
        "summary": "Create a new product",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
        "responses": { "201": { "description": "Product created" } }
      }
    },
    "/products/{id}": {
      "put": {
        "summary": "Update a product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
        "responses": { "200": { "description": "Product updated" } }
      },
      "delete": {
        "summary": "Delete a product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product deleted" } }
      }
    },

    "/orders": {
      "get": { "summary": "Get all orders", "responses": { "200": { "description": "List of orders" } } },
      "post": {
        "summary": "Create a new order",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
        "responses": { "201": { "description": "Order created" } }
      }
    },
    "/orders/{id}": {
      "put": {
        "summary": "Update an order",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
        "responses": { "200": { "description": "Order updated" } }
      },
      "delete": {
        "summary": "Delete an order",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order deleted" } }
      }
    },

    "/customers": {
      "get": { "summary": "Get all customers", "responses": { "200": { "description": "List of customers" } } },
      "post": {
        "summary": "Create a new customer",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
        "responses": { "201": { "description": "Customer created" } }
      }
    },
    "/customers/{id}": {
      "put": {
        "summary": "Update a customer",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
        "responses": { "200": { "description": "Customer updated" } }
      },
      "delete": {
        "summary": "Delete a customer",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Customer deleted" } }
      }
    },

    "/suppliers": {
      "get": { "summary": "Get all suppliers", "responses": { "200": { "description": "List of suppliers" } } },
      "post": {
        "summary": "Create a new supplier",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Supplier" } } } },
        "responses": { "201": { "description": "Supplier created" } }
      }
    },
    "/suppliers/{id}": {
      "put": {
        "summary": "Update a supplier",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Supplier" } } } },
        "responses": { "200": { "description": "Supplier updated" } }
      },
      "delete": {
        "summary": "Delete a supplier",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Supplier deleted" } }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "category": { "type": "string" },
          "inStock": { "type": "boolean" }
        },
        "required": ["name", "price", "category"]
      },
      "Order": {
        "type": "object",
        "properties": {
          "productId": { "type": "string" },
          "quantity": { "type": "number" },
          "customerName": { "type": "string" },
          "customerEmail": { "type": "string" },
          "status": { "type": "string" }
        },
        "required": ["productId", "quantity", "customerName", "customerEmail"]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" }
        },
        "required": ["name", "email"]
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" }
        },
        "required": ["name", "email"]
      }
    }
  }
}